Fix error: conversion to 'size_t' {aka 'long unsigned int'} from 'int' may change the sign of the result [-Werror=sign-conversion]

+++ link/.local/include/utf8width.hpp
@@ -56,7 +56,7 @@ utf8width(const std::string& s)
 		if (codepoint == 0)
 			break;
 
-		num_cols += utf8proc_charwidth(codepoint);
+		num_cols += static_cast<size_t>(utf8proc_charwidth(codepoint));
 		dst += bytes_read;
 	}
